Answer a questions
```
a = "SCHL"
b = "SCHL"
del a
del b
c = "SCHL"
```

How many string objects are created by the execution
of the first line of the script? (106-line1.txt)
How many string objects are created by the execution
of the second line of the script (106-line2.txt)
After the execution of line 3,
is the string object pointed by a deleted? (106-line3.txt)
After the execution of line 4,
is the string object pointed by b deleted? (106-line4.txt)
How many string objects are created by the execution
of the last line of the script (106-line5.txt)